requirements.txt
flask>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
python-dotenv>=0.19.0
jieba>=0.42.0
apscheduler>=3.10.0
